refactor(cross-engine-migration): EndpointForm을 별도 파일로 분리
대형 파일 정리 6번째(Phase F). cross_engine_migration_dialog.py(1796줄)는
CrossEngineMigrationDialog(1456줄, 상태 공유 심한 단일 거대 클래스) +
완전히 독립적인 EndpointForm(소스/타겟 연결 입력 폼, 288줄) +
CrossEngineMigrationWizard(4줄 글루)로 구성됨. EndpointForm만 분리.

- cross_engine_migration_endpoint_form.py (신규): EndpointForm
- cross_engine_migration_dialog.py: 1796 -> 1508줄 (재수입)

검증: 관련 pytest 64 passed (최초 실행에서 make_connection_payload import
누락으로 13개 실패 확인 → 수정 후 전부 통과). 전체 pytest 1809 passed /
1 known-flaky 제외. py_compile 통과. CrossEngineMigrationDialog 오프스크린
오픈 확인. 앱 스모크 성공.
